Update transcript metadatda for tximix() imported data

This function expects a SummarizedExperiment as output by tximix(), and if possible, it will update the metadata on the transcripts (rowData and/or rowRanges depending on the value of ranges), using metadata where the index digest matches those in tximeta locations. Additionally, GRanges or data.frame-type data can be provided directly to txpData, although this is not a persistent method for linking data to metadata. See tximixInspectDigests() for information on ascertaining which sources are present, and how to link data to local metadata.

tximixUpdateTxpData(
  se,
  txpData = NULL,
  ranges = FALSE,
  order = c("annotated", "novel"),
  key = "tx_name"
)

Arguments

se

the SummarizedExperiment (SE) output by tximix()

txpData

either GRanges or data.frame-type object to use if there is not a match based on digest. This is used on a one-time basis, and transcripts will be marked in metadata columns as index = "user"``. See linkedTxomeorlinkedTxpData` for persistent metadata storage/retrieval

ranges

logical, whether to add rowRanges (or just rowData)

order

order in which to update the metadata, by default annotation then novel

key

the name of the column to use as the key for merging metadata with rownames(se). Defaults to key="tx_name" which often matches the transcript names in GENCODE

Value

a SummarizedExperiment with additional rowData, or a RangedSummarizedExperiment with additional ranges/data

Examples

if (FALSE) { # \dontrun{
example(tximix)

# build custom novel GRanges data
novel <- data.frame(
  seqnames = paste0("chr", rep(1:22, each=500)),  
  start = 1e6 + 1 + 0:499 * 1000, end = 1e6 + 1 + 0:499 * 1000 + 1000 - 1,
  strand = "+", tx_name = paste0("novel", 1:(22*500)),
  gene_id = paste0("novel_gene", rep(1:(22*10), each=50)), type = "protein_coding"
)
novel_gr <- as(novel, "GRanges")
names(novel_gr) <- novel$tx_name
seqinfo(novel_gr) <- seqinfo(se) # needs to have consistent seqinfo with `se`

# now update the metadata + ranges:
se_with_ranges <- tximixUpdateTxpData(
  se, novel_gr, ranges=TRUE
)
mcols(se_with_ranges)
} # }